requirements.txt
spotipy==2.23.0
ytmusicapi==1.5.2
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
pyopenssl==24.0.0
